+++ src/components/organisms/core/Footer/styled/Sitemap.jsx
@@ -1,0 +1,33 @@
+/*
+ * Copyright (C) 2018-present Arctic Ice Studio <development@arcticicestudio.com>
+ * Copyright (C) 2018-present Sven Greb <development@svengreb.de>
+ *
+ * Project:    Nord Docs
+ * Repository: https://github.com/arcticicestudio/nord-docs
+ * License:    MIT
+ */
+
+import styled from "styled-components";
+
+/**
+ * A container for sitemap components.
+ *
+ * @author Arctic Ice Studio <development@arcticicestudio.com>
+ * @author Sven Greb <development@svengreb.de>
+ * @since 0.5.0
+ */
+const Sitemap = styled.div`
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  margin-top: 1em;
+  margin-bottom: 1em;
+
+  ${({ theme }) => theme.media.tabletPortrait`
+    align-items: unset;
+    margin-top: 0;
+    margin-bottom: 0;
+  `};
+`;
+
+export default Sitemap;